### 2018-04-18 #683 @whedon
[PRE REVIEW]: Sentiment Analysis of Twitter Data (SAoTD)
**Submitting author:** @evan-l-munson (<a href="http://orcid.org/0000-0002-9958-6800">Evan Munson</a>)
**Repository:** <a href="https://github.com/evan-l-munson/SAoTD" target ="_blank">https://github.com/evan-l-munson/SAoTD</a>
**Version:** 0.1.0
