### 2018-04-18 #683 @arfon
Hi @evan-l-munson - please review our [submission guidelines](https://joss.theoj.org/about#author_guidelines) and prepare a paper to accompany this submission.

Any questions, don't hesitate to ask.
