+++ R/stop_words.R
@@ -1,14 +1,0 @@
-#' @title Stop Words
-#' 
-#' @description 
-#' Stop Words Dictionary
-#' 
-#' @usage data(stop_words)
-#' 
-#' @format A \code{tribble} with 1,149 rows and 1 variables
-#' 
-#' \describe{
-#'  \item{word}{List of stop words.}
-#'  }
-#'
-"stop_words"
