+++ R/Bing.R
@@ -1,15 +1,0 @@
-#' @title Bing Lexicon Dictionary
-#' 
-#' @description 
-#' Dataset containing the Bing Sentiment Lexicon Dictionary
-#' 
-#' @usage data(Bing)
-#' 
-#' @format A \code{tribble} with 6,788 rows and 2 variables
-#' 
-#' \describe{
-#'  \item{Token}{A word}
-#'  \item{Sentiment}{The determined sentiment of a Token.  Either "positive" or "negative".}
-#'  }
-#'
-"Bing"
