### 2018-05-11 f21c5db
paper.md added for JOSS submission
