+++ paper.bib
@@ -1,0 +1,48 @@
+@inproceedings{Hu2004,
+	title={{Mining and summarizing customer reviews}},
+  author={Hu, Minqing and Liu, Bing},
+  booktitle={Proceedings of the tenth ACM SIGKDD international conference on Knowledge discovery and data mining},
+  pages={168--177},
+  year={2004},
+  organization={ACM}
+}
+
+@Manual{Wickham2017,
+    title = {tidyverse: Easily Install and Load the 'Tidyverse'},
+    author = {Hadley Wickham},
+    year = {2017},
+    note = {R package version 1.2.1},
+    url = {https://CRAN.R-project.org/package=tidyverse},
+  }
+  
+  @book{Silge2017,
+	address = {Sebastopol},
+	author = {Silge, Julia and Robinson, David},
+	edition = {1},
+	editor = {Tache, Nicole},
+	isbn = {978-1-491-98165-8},
+	pages = {178},
+	publisher = {O'Reilly},
+	title = {{Text Mining with R a Tidy Approach}},
+	url = {http://tidytextmining.com/},
+	year = {2017}
+}
+
+@Manual{Nikita2016,
+  title = {ldatuning: Tuning of the Latent Dirichlet Allocation Models Parameters},
+  author = {Murzintcev Nikita},
+  year = {2016},
+  note = {R package version 0.2.0},
+  url = {https://CRAN.R-project.org/package=ldatuning},
+}
+
+@article{Grun2011,
+  title = {{topicmodels}: An {R} Package for Fitting Topic Models},
+  author = {Bettina Gr\"un and Kurt Hornik},
+  journal = {Journal of Statistical Software},
+  year = {2011},
+  volume = {40},
+  number = {13},
+  pages = {1--30},
+  doi = {10.18637/jss.v040.i13},
+}
