+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: SAoTD
 Type: Package
 Title: Sentiment Analysis of Twitter Data
-Version: 1.0
+Version: 1.0.0
 Date: 2018-04-17
 Author: Evan L. Munson 
 Maintainer: Evan L. Munson <evan.l.munson@gmail.com>
