### 2018-05-19 9910848
Delete NAMESPACE
