+++ NAMESPACE
@@ -1,73 +1,0 @@
-# Generated by roxygen2: do not edit by hand
-
-export("%>%")
-export(Acquire)
-export(Bigram)
-export(Bigram.Network)
-export(BoxPlot)
-export(Corups.Distribution)
-export(Distribution)
-export(Max.Scores)
-export(Merge.Terms)
-export(Min.Scores)
-export(Number.Topics)
-export(PosNeg.Words)
-export(Scores)
-export(Tidy)
-export(TimeScale)
-export(Trigram)
-export(Tweet.Topics)
-export(Unigram)
-export(ViolinPlot)
-export(Word.Corr)
-export(Word.Corr.Plot)
-export(WorldMap)
-export(run_my_app)
-import(ggplot2)
-import(stringi)
-importFrom(Rcpp,evalCpp)
-importFrom(dplyr,anti_join)
-importFrom(dplyr,arrange)
-importFrom(dplyr,count)
-importFrom(dplyr,desc)
-importFrom(dplyr,distinct)
-importFrom(dplyr,filter)
-importFrom(dplyr,group_by)
-importFrom(dplyr,inner_join)
-importFrom(dplyr,mutate)
-importFrom(dplyr,quo)
-importFrom(dplyr,select)
-importFrom(dplyr,summarize)
-importFrom(dplyr,top_n)
-importFrom(dplyr,transmute)
-importFrom(dplyr,ungroup)
-importFrom(ggraph,geom_edge_link)
-importFrom(ggraph,geom_node_point)
-importFrom(ggraph,geom_node_text)
-importFrom(ggraph,ggraph)
-importFrom(igraph,graph_from_data_frame)
-importFrom(ldatuning,FindTopicsNumber)
-importFrom(lubridate,as_date)
-importFrom(magrittr,"%>%")
-importFrom(plyr,rename)
-importFrom(purrr,map_df)
-importFrom(reshape2,melt)
-importFrom(scales,rescale)
-importFrom(shiny,shinyAppDir)
-importFrom(stats,median)
-importFrom(stats,reorder)
-importFrom(stringr,str_replace_all)
-importFrom(tidyr,separate)
-importFrom(tidyr,spread)
-importFrom(tidytext,cast_dtm)
-importFrom(tidytext,get_sentiments)
-importFrom(tidytext,unnest_tokens)
-importFrom(topicmodels,LDA)
-importFrom(topicmodels,terms)
-importFrom(topicmodels,topics)
-importFrom(twitteR,searchTwitter)
-importFrom(twitteR,setup_twitter_oauth)
-importFrom(twitteR,twListToDF)
-importFrom(utils,head)
-importFrom(widyr,pairwise_cor)
-useDynLib(SAoTD, .registration = TRUE)
