### 2018-06-04 #683 @leeper
@kbenoit Would you be willing to review this submission for JOSS?
