### 2018-06-04 #764 @whedon
[REVIEW]: Sentiment Analysis of Twitter Data (SAoTD)
**Submitting author:** @evan-l-munson (<a href="http://orcid.org/0000-0002-9958-6800">Evan Munson</a>)
**Repository:** <a href="https://github.com/evan-l-munson/SAoTD" target ="_blank">https://github.com/evan-l-munson/SAoTD</a>
**Version:** 0.2.0

    No, but many R packages using standard licenses only name the license in the `DESCRIPTION` file.  The `README.md` refers to "All code is licensed GL", this should be "GPL-x".
    No: GitHub release is 6 commits behind the master.  Master has v1.0.0.
    Yes, apparently all.
- [x] Does the full list of paper authors seem appropriate and complete?
    More than complete, as the 2nd through last author have an unknown contribution to the software.  They have not made any GitHub commits, and there is no record of them having authored any of the functions (for instance via `@author`).

    No, the paper focuses more on what the package does than providing an explicit statement of need.  This could be made more clear (and at little cost or effort).
