### 2018-06-07 #764 @kbenoit
No, but many R packages using standard licenses only name the license in the `DESCRIPTION` file.  The `README.md` refers to "All code is licensed GL", this should be "GPL-x".
    No: GitHub release is 6 commits behind the master.  Master has v1.0.0.
    Yes, apparently all.
- [x] Does the full list of paper authors seem appropriate and complete?
    More than complete, as the 2nd through last author have an unknown contribution to the software.  They have not made any GitHub commits, and there is no record of them having authored any of the functions (for instance via `@author`).

    No, the paper focuses more on what the package does than providing an explicit statement of need.  This could be made more clear (and at little cost or effort).

### Comments

#### On the paper

This paper describes an R package that provides a workflow for analyzing sentiment and topics in Twitter text, wrapping around packages such as **twitteR**, **tidytext**, and **topicmodels**.  The package contains a number of useful analytic functions for looking at Twitter data, and these are clearly demonstrated in the vignette.  Some of these could be useful to other forms of text, but the package is specially designed to work with Twitter data, including not just the import of this data but also working with Twitter-specific handles such as hashtags and usernames

This paper makes a nice, short article that should be published, but could be improved by addressing a few relatively minor issues:

-  A few sentences at the outset describing the ideal use of the package and how it fits into other text analysis workflows using R would be welcome.
-  The package presumes a knowledge of the tidyverse approach to R, which is fine, but the paper (and package documentation) could mention this more explicitly.
-  Should the lexicon used for sentiment analysis be called the "Bing lexicon" or the Hu and Liu (2004) sentiment dictionary?
-  "Twitter" and "Tweet" should be capitalized in the text.

#### On the package

The package works, and I have seen much worse source code in widely-used R packages published on CRAN.  These are some suggestions for improving the code and the package, not necessarily linked to the paper and whether it should be published.  (I leave to the editor to decide.)

_Naming_.  This is a matter of preference, although there are some [emerging guidelines](http://r-pkgs.had.co.nz/style.html#style-notnam) designed to reduce the [chaos in the R world](https://www.r-bloggers.com/consistent-naming-conventions-in-r/).  This paper combines capitalized object names with lower-cased object names, and function names with `.` (e.g. e.g. `PosNeg.Words()`), which is generally discouraged due to the indeterminacy with the S3 dispatch system.  Do the functions really need to be capitalized?  Also naming is not consistent: `Word.Corr()` is "object.verb" but `Number.Topics` is verb.object.

The package name itself runs contrary to [this advice](http://r-pkgs.had.co.nz/package.html#naming) from Hadley Wickham:
> Avoid using both upper and lower case letters: doing so makes the package name hard to type and even harder to remember.

_Unnecessary C++ code._  Why is there a function `rcpp_hello_world()`?  This looks like demonstration code that should be removed.

_Data copyright issues._  Can distribute the data in `raw_tweets`.  That material is copyright of the authors of the Tweets.  At the least, it may require some attribution.  It would be worth reviewing the Twitter terms of service about this.

_Data object loading_.  Set `LazyLoad: true` in `DESCRIPTION` else code like the (not run) examples for `Scores()` will not work.

_(non-)Object orientation_.  None of the functions use generics and method dispatch, but rather check the class of the input objects using conditionals within each function (e.g. [here](https://github.com/evan-l-munson/SAoTD/blob/master/R/Functions.R#L106-L108).  This makes extending the package harder, in addition to being more error-prone.   The function names are very generic, furthermore, such as `BoxPlot()`, or `Tidy()`.  Other packages have functions named `tidy()`, but they are defined for specific object classes.  I suggest using more distinctive names to differentiate this package's functions from those found in other packages, and/or method dispatch for specific object classes.  Simplyu capitalizing the function names is likely to confuse some users.

_Code organization_.  Nearly all functions are in a single long .R file called `Function.R`.  Splitting this into smaller files would make the code organization clearer.

_Examples_.  Most examples are not run, due to the difficulties of connecting to the Twitter API using authentication.  But this is not true for an functions that only use `raw_tweets`, such as `Bigram()`, `Bigram.Network()`, `BoxPlot()`, etc.  Furthermore, this code does not run as written, because `raw_tweets` is not lazy loaded.  In addition, there is no need to load the package in the examples (using `library(SAoTD)`) because the help functions should be only accessible if the package is already loaded.

_Tests_.  The file `tests/testthat/test_Acquire.R` contains Twitter authentication keys.  These should be removed (and changed, since they will remain visible in the git history).
