### 2018-06-07 #764 @leeper
Excellent review, @kbenoit! Thank so much!

@evan-l-munson Can you address the issues raised in the review - particularly the missing checked items from the review checklist and the other useful suggestions raised in the review?
