### 2018-06-07 #764 @evan-l-munson
Thank you for the review.

I will work on those notes/corrections as soon as I get a chance (just moved my family across the United States and started a new job).

Evan Munson

Sent from my iPhone

> On Jun 7, 2018, at 01:15, Thomas J. Leeper <notifications@github.com> wrote:
>
> Excellent review, @kbenoit! Thank so much!
>
> @evan-l-munson Can you address the issues raised in the review - particularly the missing checked items from the review checklist and the other useful suggestions raised in the review?
>
