### 2018-10-22 7b44d98
JOSS package comments (more work to do)
