+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: "Sentiment Analysis of Twitter Data (SAoTD)"
+title: "Sentiment Analysis of Twitter Data (saotd)"
 authors:
 - affiliation: 1
   name: Evan L. Munson
@@ -27,9 +27,9 @@ affiliations:
   name: Air Force Institure of Technology
 ---
 
-SAoTD is an R interface to the Twitter API and can be used to acquire tweets based on user selected \#hashtags.  The package will clean and tidy the Twitter data, determine the latent topics within the tweets utilizing Latent Dirichlet Allocation (LDA), determine a sentiment score using the Bing lexicon dictionary and output visualizations.
+saotd is an R interface to the Twitter API and can be used to acquire tweets based on user selected \#hashtags.  The package will clean and tidy the Twitter data, determine the latent topics within the tweets utilizing Latent Dirichlet Allocation (LDA), determine a sentiment score using the Bing lexicon dictionary and output visualizations.
 
-The package is available on [GitHub](https://github.com/evan-l-munson/SAoTD) and archived on [Zenodo](https://zenodo.org/record/1219852#.WtdNuchryfd).  The package was developed with the intention of the user creating a personal [Twitter](hjttps://twitter.com) account which will then allow  a user to access the twitter API throught the [Twitter Developers Account](https://dev.twitter.com/) site.  Once a user has access to the Twitter Developers Account they will have the ability to create an application which will then provide the user with access tokens.  These access tokens will then allow the user to begin acquiring tweets usint eh SAoTD package.
+The package is available on [GitHub](https://github.com/evan-l-munson/saotd) and archived on [Zenodo](https://zenodo.org/record/1219852#.WtdNuchryfd).  The package was developed with the intention of the user creating a personal [Twitter](hjttps://twitter.com) account which will then allow  a user to access the twitter API throught the [Twitter Developers Account](https://dev.twitter.com/) site.  Once a user has access to the Twitter Developers Account they will have the ability to create an application which will then provide the user with access tokens.  These access tokens will then allow the user to begin acquiring tweets usint eh saotd package.
 
 The package is laid out in five different categories: Acquire, Explore, Topic Analysis, Sentiment Calculation, and Visualizations.
 
@@ -41,7 +41,7 @@ The package is laid out in five different categories: Acquire, Explore, Topic An
 
 The package utilizes tidy dataferames and therefore depends on the tidyverse package [@Wickham2017] and additionally uses the tidytext package [@Silge2017].  The number of latent topics is determined using the ldatuning package [@Nikita2016] and the latent dirichlet allocation (LDA) topics is determined using the topicmodels package [@Grun2011].
 
-The SAoTD package has research applications in many disciplines which intend to use twitter text and sentiment analysis.  The package was created to quickly determine the sentiment of twitter and to inform analysts on the opinions contained within tweets.
+The saotd package has research applications in many disciplines which intend to use twitter text and sentiment analysis.  The package was created to quickly determine the sentiment of twitter and to inform analysts on the opinions contained within tweets.
 
 # References
 
