+++ tests/testthat/test_BigramNetwork.R
@@ -5,7 +5,7 @@ text <- "This is the website for “R for Data Science”.
 You’ll learn how to get your data into R, get it into the most useful structure, transform it, visualise it and model it." 
 test_text_df <- as.data.frame(x = text)
 
-test_bigram_df <- SAoTD::Bigram(DataFrame = test_text_df)
+test_bigram_df <- saotd::Bigram(DataFrame = test_text_df)
 
 incorrect_bigram_df <- dplyr::tribble(
   ~word, ~word2, ~n,
@@ -14,14 +14,14 @@ incorrect_bigram_df <- dplyr::tribble(
   "youll", "learn", as.integer(1)
 )
 
-p <- SAoTD::Bigram.Network(BiGramDataFrame = test_bigram_df, number = 1)
+p <- saotd::Bigram.Network(BiGramDataFrame = test_bigram_df, number = 1)
 
 # Tests
 test_that("The Bigram.Network function is working as properly", {
 
-  expect_error(object = SAoTD::Bigram.Network(BiGramDataFrame = text), "The input for this function is a Bigram data frame.")
-  expect_error(object = SAoTD::Bigram.Network(BiGramDataFrame = test_bigram_df, number = 0), "You must choose number of Bi-Grams greater than 1.")
-  expect_error(object = SAoTD::Bigram.Network(BiGramDataFrame = incorrect_bigram_df), "The data frame is not properly constructed.  The data frame must have three columns: word1, word2 and n.")
+  expect_error(object = saotd::Bigram.Network(BiGramDataFrame = text), "The input for this function is a Bigram data frame.")
+  expect_error(object = saotd::Bigram.Network(BiGramDataFrame = test_bigram_df, number = 0), "You must choose number of Bi-Grams greater than 1.")
+  expect_error(object = saotd::Bigram.Network(BiGramDataFrame = incorrect_bigram_df), "The data frame is not properly constructed.  The data frame must have three columns: word1, word2 and n.")
   
 })
 
