+++ inst/doc/SAoTD.R
@@ -1,207 +1,0 @@
-## ----warning=FALSE, message=FALSE----------------------------------------
-library(SAoTD)
-library(tidytext)
-library(dplyr)
-library(stringr)
-library(knitr)
-library(utils)
-
-## ----eval=FALSE----------------------------------------------------------
-#  consumer_key <- "XXXXXXXXXXXXXXXXXXXXXXXXX"
-#  consumer_secret <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-#  access_token <- "XXXXXXXXXXXXXXXXXX-XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-#  access_secret <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-#  
-#  hashtags <- c("#job", "#Friday", "#fail", "#icecream", "#random", "#kitten", "#airline")
-#  
-#  Acquire(consumer_key = consumer_key,
-#          consumer_secret = consumer_secret,
-#          access_token = access_token,
-#          access_secret = access_secret,
-#          HT = hashtags,
-#          num_tweets = 1000,
-#          file_name = "test_tweets.RData",
-#          distinct = TRUE)
-#  
-#  load("test_tweets.RData")
-
-## ------------------------------------------------------------------------
-data("raw_tweets")
-TD <- raw_tweets
-
-## ------------------------------------------------------------------------
-TD_Tidy <- SAoTD::Tidy(DataFrame = TD)
-
-TD_Tidy$Token[11:20] %>% 
-  knitr::kable("html")
-
-## ----message=FALSE-------------------------------------------------------
-SAoTD::Unigram(DataFrame = TD) %>% 
-  dplyr::top_n(10) %>% 
-  knitr::kable("html", caption = "Twitter data Uni-Grams")
-
-## ----message=FALSE-------------------------------------------------------
-SAoTD::Bigram(DataFrame = TD) %>% 
-  dplyr::top_n(10) %>% 
-  knitr::kable("html", caption = "Twitter data Bi-Grams")
-
-## ----message=FALSE-------------------------------------------------------
-SAoTD::Trigram(DataFrame = TD) %>% 
-  dplyr::top_n(10) %>% 
-  knitr::kable("html", caption = "Twitter data Tri-Grams")
-
-## ----message=FALSE, message=FALSE, error=FALSE---------------------------
-TD_Merge <- Merge.Terms(DataFrame = TD, term = "ice cream", term_replacement = "icecream")
-
-## ----message=FALSE-------------------------------------------------------
-SAoTD::Unigram(DataFrame = TD_Merge) %>% 
-  dplyr::top_n(10) %>% 
-  knitr::kable("html", caption = "Twitter data Uni-Grams")
-
-## ----message=FALSE-------------------------------------------------------
-SAoTD::Bigram(DataFrame = TD_Merge) %>% 
-  dplyr::top_n(10) %>% 
-  knitr::kable("html", caption = "Twitter data Bi-Grams")
-
-## ----message=FALSE-------------------------------------------------------
-SAoTD::Trigram(DataFrame = TD_Merge) %>% 
-  dplyr::top_n(10) %>% 
-  knitr::kable("html", caption = "Twitter data Tri-Grams")
-
-## ----fig.align='center'--------------------------------------------------
-TD_Corr <- SAoTD::Word.Corr(DataFrameTidy = TD_Tidy, 
-                            number = 200, 
-                            sort = TRUE)
-
-SAoTD::Word.Corr.Plot(WordCorr = TD_Corr, 
-                      Correlation = .1, 
-                      layout = "fr", 
[…]
