### 2018-10-28 ee0a4c4
corrected to 0 notes, but 3 warnings in test()
