+++ man/SAoTD.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/SAoTD.R
+% Please edit documentation in R/saotd.R
 \docType{package}
-\name{SAoTD}
-\alias{SAoTD}
-\alias{SAoTD-package}
+\name{saotd}
+\alias{saotd}
+\alias{saotd-package}
 \title{Sentiment Analysis of Twitter Data}
 \description{
 This analytic is an in initial forray into sentiment analysis.  
@@ -12,6 +12,6 @@ ingest tweets of their interest, clean / tidy data, perform topic modeling if in
 scores utilizing the Bing Lexicon, and output visualizations.
 
 For additional information and a description on how to use the package, 
-please look at the package vignette: utils::vignette('SAoTD').
+please look at the package vignette: utils::vignette('saotd').
 }
 \keyword{internal}
