+++ NAMESPACE
@@ -68,4 +68,4 @@ importFrom(twitteR,setup_twitter_oauth)
 importFrom(twitteR,twListToDF)
 importFrom(utils,head)
 importFrom(widyr,pairwise_cor)
-useDynLib(SAoTD, .registration = TRUE)
+useDynLib(saotd, .registration = TRUE)
