### 2018-10-28 f1c1048
add Rmpfr to description, maybe clear travis error
