### 2018-10-28 ed702f2
Rmpfr adjustment??
