+++ .travis.yml
@@ -25,7 +25,6 @@ apt_packages:
  
 r_packages:
  - covr
- - Rmpfr
-
+ 
 after_success:
  - Rscript -e 'library(covr); codecov()'
