### 2018-11-04 d8be00f
added build_vignettes = TRUE
