+++ DESCRIPTION
@@ -41,3 +41,4 @@ SystemRequirements: GSL (>=2.4),
     MPFR (>= 4.0.0), 
     udunits2 (>=2.2.26-3)
 Encoding: UTF-8
+LazyLoad: true
