+++ R/Functions.R
@@ -1,7 +1,7 @@
 
 
 
-# Acquire -----------------------------------------------------------------
+# tweet_acquire -----------------------------------------------------------------
 
 #' @title Acquire Twitter Tweets  
 #'
@@ -34,7 +34,7 @@
 #' 
 #' hashtags <- c("#job", "#Friday", "#fail", "#icecream", "#random", "#kitten", "#airline")
 #' 
-#' Acquire(consumer_key = consumer_key, 
+#' tweet_acquire(consumer_key = consumer_key, 
 #'         consumer_secret = consumer_secret, 
 #'         access_token = access_token, 
 #'         access_secret = access_secret, 
@@ -47,7 +47,7 @@
 #' }
 #' @export 
 
-Acquire <- function(consumer_key, consumer_secret, access_token, access_secret, HT, num_tweets, file_name, distinct = TRUE) {
+tweet_acquire <- function(consumer_key, consumer_secret, access_token, access_secret, HT, num_tweets, file_name, distinct = TRUE) {
   
   options(httr_oauth_cache = TRUE)
   
@@ -95,12 +95,12 @@ Acquire <- function(consumer_key, consumer_secret, access_token, access_secret, 
 #' \dontrun{
 #' library(saotd)
 #' data <- raw_tweets
-#' tidy_data <- Tidy(DataFrame = data)
+#' tidy_data <- tweet_tidy(DataFrame = data)
 #' tidy_data
 #' }
 #' @export
 
-Tidy <- function(DataFrame) {
+tweet_tidy <- function(DataFrame) {
   
   if(!is.data.frame(DataFrame)) {
     stop('The input for this function is a data frame.')
@@ -142,14 +142,14 @@ Tidy <- function(DataFrame) {
 #' \dontrun{
 #' library(saotd)
 #' data <- raw_tweets
-#' data <- Merge.Terms(DataFrame = data, 
+#' data <- merge_terms(DataFrame = data, 
 #'                     term = "ice cream", 
 #'                     term_replacement = "ice_cream")
 #' data 
 #' }
 #' @export
 
-Merge.Terms <- function(DataFrame, term, term_replacement){
+merge_terms <- function(DataFrame, term, term_replacement){
   
   if(!is.data.frame(DataFrame)) {
     stop('The input for this function is a data frame.')
@@ -180,12 +180,12 @@ Merge.Terms <- function(DataFrame, term, term_replacement){
 #' \dontrun{
 #' library(saotd)
 #' data <- raw_tweets
-#' TD_Unigram <- Unigram(DataFrame = data)
+#' TD_Unigram <- unigram(DataFrame = data)
 #' TD_Unigram
 #' }             
 #' @export
 
-Unigram <- function(DataFrame){
+unigram <- function(DataFrame){
   
   if(!is.data.frame(DataFrame)) {
     stop('The input for this function is a data frame.')
@@ -224,12 +224,12 @@ Unigram <- function(DataFrame){
 #' \dontrun{
 #' library(saotd)
 #' data <- raw_tweets
-#' TD_Bigram <- Bigram(DataFrame = data)
+#' TD_Bigram <- bigram(DataFrame = data)
 #' TD_Bigram
 #' }                
 #' @export
 
-Bigram <- function(DataFrame){
+bigram <- function(DataFrame){
   
   if(!is.data.frame(DataFrame)) {
     stop('The input for this function is a data frame.')
@@ -273,12 +273,12 @@ Bigram <- function(DataFrame){
 #' \dontrun{
 #' library(saotd)
 #' data <- raw_tweets
-#' TD_Trigram <- Trigram(DataFrame = data)
+#' TD_Trigram <- trigram(DataFrame = data)
 #' TD_Trigram
 #' }
 #' @export
 
-Trigram <- function(DataFrame) {
[…]
