### 2018-11-07 b411324
test trigrams
