+++ tests/testthat/test_Trigram.R
@@ -1,6 +1,6 @@
 
 text <- "Trigrams provide significant awesome information pertaining to the world up down around them.  
-  Awesome information, empowers the world to make everything better."
+Awesome information, empowers the world to make everything better."
 test_trigram_df <- as.data.frame(x = text)
 
 
