+++ tests/testthat/test_Trigram.R
@@ -1,3 +1,4 @@
+testthat::context("trigram")
 
 text <- "Trigrams provide significant awesome information pertaining to the world up down around them.  
 Awesome information, empowers the world to make everything better."
