+++ R/number_topics.R
@@ -1,0 +1,93 @@
+
+#' @title Number Topics
+#'
+#' @description Determines the optimal number of Latent topics within a dataframe by tuning the Latent Dirichlet Allocation (LDA) model parameters.  
+#' Uses the `ldatuning` package and outputs an ldatuning plot.  __This process can be time consuming depending on the size of the input dataframe.__
+#'
+#' @param DataFrame DataFrame of Twitter Data.
+#' @param num_cores The number of CPU cores to processes models simultaneously (2L for dual core processor).
+#' @param min_clusters Lower range for the number of clusters.
+#' @param max_clusters Upper range for the number of clusters.
+#' @param skip Integer; The number of clusters to skip between entries.
+#' @param set_seed Seed for reproducable results.
+#' 
+#' @importFrom dplyr mutate group_by count anti_join quo
+#' @importFrom stringr str_replace_all
+#' @importFrom tidytext unnest_tokens cast_dtm 
+#' @importFrom ldatuning FindTopicsNumber
+#' @importFrom scales rescale
+#' @importFrom reshape2 melt
+#' @import ggplot2
+#' 
+#' @return A Tidy DataFrame.
+#' 
+#' @examples 
+#' \dontrun{
+#' library(saotd)
+#' data <- raw_tweets
+#' LDA_Topic_Plot <- number_topics(DataFrame = data,
+#'                                 num_cores = 2L,
+#'                                 min_clusters = 2,
+#'                                 max_clusters = 12, 
+#'                                 skip = 2,
+#'                                 set_seed = 1234)
+#'
+#' LDA_Topic_Plot 
+#' }
+#' @export
+
+number_topics <- function(DataFrame, num_cores, min_clusters = 2, max_clusters = 12, skip = 2, set_seed = 1234) {
+  
+  if(!is.data.frame(DataFrame)) {
+    stop('The input for this function is a data frame.')
+  }
+  
+  set_seed
+  
+  text <- dplyr::quo(text)
+  key <- dplyr::quo(key)
+  word <- dplyr::quo(word)
+  n <- dplyr::quo(n)
+  
+  lda_prep <- DataFrame %>% 
+    dplyr::mutate(text = base::iconv(DataFrame$text, "latin1", "ASCII", sub="")) %>% 
+    dplyr::mutate(text = stringr::str_replace_all(text, "#", "")) %>% # Remove hashtag
+    dplyr::mutate(text = stringr::str_replace_all(text, "[:punct:]", "")) %>% # Remove punctuation
+    dplyr::mutate(text = stringr::str_replace_all(text, "RT", "")) %>% # Remove retweet note
+    dplyr::mutate(text = stringr::str_replace_all(text, "&amp", "")) %>% # Remove Accelerated Mobile Pages (AMP) note
+    dplyr::mutate(text = stringr::str_replace_all(text, "https://t.co/[A-Za-z\\d]+|http://[A-Za-z\\d]+|&amp;|&lt;|&gt;|RT|https", "")) %>%  # Remove links
+    dplyr::group_by(key) %>%
+    tidytext::unnest_tokens(word, text) %>% 
+    dplyr::anti_join(tidytext::stop_words) %>% 
+    dplyr::count(key, word, sort = TRUE) %>% 
+    tidytext::cast_dtm(key, word, n) # create DTM
+  
+  # Compute Values
+  values <- ldatuning::FindTopicsNumber(lda_prep, 
+                                        topics = seq(from = min_clusters, to = max_clusters, by = skip),
+                                        metrics = c("Griffiths2004", "CaoJuan2009", "Arun2010", "Deveaud2014"),
+                                        method = "Gibbs",
+                                        mc.cores = num_cores,
+                                        verbose = TRUE)
+  
+  # Plot
[…]
