### 2018-11-19 5802963
added new example data from kaggle competition
