### 2018-11-19 #764 @evan-l-munson
Good evening @arfon, I think I made all the requested corrections.  Please let me know if you need anything else corrected/adjusted.  Thanks for the patience and assistance in this process.
