### 2018-12-10 #764 @kbenoit
I've had a chance to look at the package again, and I am pleased to report that it and the paper are much improved. @evan-l-munson has done a very good job of addressing my concerns [above](https://github.com/openjournals/joss-reviews/issues/764#issuecomment-395321533), if not as good a job of summarizing in a PR or memo what these changes were 😉 .

_Package_.  The code organization is much better (and will be easier to maintain or for potential contributors to absorb). I also like the function naming much better than before - the function index looks more tidy and sensible now.

There is still unnecessary C++ code in `/src`, which creates the function `rcpp_hello_world()` in the package index. This should simply be removed.

_Paper._ The paper does a much better job now of explaining the package and its purpose, and the vignette rounds this out nicely.

Subject to the Hello World change, :+1:.
