### 2018-12-10 #764 @arfon
Thanks @kbenoit. @evan-l-munson - please make these final changes to your package and we can move forward accepting this submission.
