### 2018-12-13 #764 @evan-l-munson
Gentlemen, I missed your email by accident.  I appreciate these comments and will try to get them corrected/accounted for this weekend or sometime during the holidays.  Thank you!
