+++ R/RcppExports.R
@@ -1,7 +1,7 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-rcpp_hello_world <- function() {
-    .Call(`_saotd_rcpp_hello_world`)
+rcpp_saotd <- function() {
+    .Call(`_saotd_rcpp_saotd`)
 }
 
