### 2018-12-16 2a14738
vignette corrections
