+++ vignettes/SAoTD.Rmd
@@ -65,7 +65,7 @@ However is you want to acquire your own tweets, you will first have to:
     
 With these steps complete you now have access to the twitter API.
 
-To acquire your own dataset of tweets you can use the `saotd::tweet_acquire` function and insert your consumer key, consumer secret key, access token and access secret key gained from the [Twitter Developers](https://dev.twitter.com/apps/) page.
+To acquire your own dataset of tweets you can use the `saotd::tweet_acquire` function and insert your consumer key, consumer secret key, access token and access secret key gained from the [Twitter Developers](https://dev.twitter.com/apps/) page.  You will also need to select the #hashtags you are interested in and the numer of tweets requested per #hashtag.
 
 ```{r eval=FALSE}
 consumer_key <- "XXXXXXXXXXXXXXXXXXXXXXXXX"
@@ -89,7 +89,8 @@ load("test_tweets.RData")
 
 ## Explore
 
-You can acquire your own data or use the dataset included with the package.  We will be using the included data `raw_tweets`.
+You can acquire your own data or use the dataset included with the package.  We will be using the included data `raw_tweets`.  This dataset was acquired from a [Twitter US Airline Sentiment](https://www.kaggle.com/crowdflower/twitter-airline-sentimen) Kaggle competition, from December 2017.  The dataset contains 14,487 tweets from 6 different hashtags (2,604 x #American, 2,220 x #Delta, 2,420 x #Southwest, 3,822 x #United, 2,913 x #US Airways, 504 x #Virgin America). 
+
 
 ```{r}
 data("raw_tweets")
@@ -101,7 +102,7 @@ The second tweet of the dataset is:  __`r TD$text[2]`__, and when it is cleaned 
 ```{r}
 TD_Tidy <- saotd::tweet_tidy(DataFrame = TD)
 
-TD_Tidy$Token[11:20] %>% 
+TD_Tidy$Token[3:8] %>% 
   knitr::kable("html")
 ```
 
@@ -127,10 +128,10 @@ saotd::trigram(DataFrame = TD) %>%
   knitr::kable("html", caption = "Twitter data Tri-Grams")
 ```
 
-Now that we have the Uni-Grams we can see that ice and cream are referring to ice cream and may be good set of words to merge into a single term.  Additionally, pet and pets could also be merged to observe more uniqueness in the data.
+Now that we have the Uni-Grams we can see that cancelled and fligt are referring to cancelled flight and may be good set of words to merge into a single term.  Additionally, pet and pets could also be merged to observe more uniqueness in the data.
 
 ```{r message=FALSE, message=FALSE, error=FALSE}
-TD_Merge <- merge_terms(DataFrame = TD, term = "ice cream", term_replacement = "icecream")
+TD_Merge <- merge_terms(DataFrame = TD, term = "cancelled flight", term_replacement = "cancelled_flight")
 ```
 
 Now that the terms have been merged, the new N-Grams are re-computed.
@@ -241,7 +242,7 @@ saotd::number_topics(DataFrame = TD,
                      set_seed = 1234)
 ```
 
-The number of topics plot shows that between 5 and 7 latent topics reside within the dataset.  Remember that the twitter data was acquired by selecting seven different hashtags `r hashtags <- c("#job", "#Friday", "#fail", "#icecream", "#random", "#kitten", "#airline")`.  We could easily see that in this case selecting 7 topics could be a correct number of topics.  However if we knew that the data contained 5 topics we could continue the analysis.  
+The number of topics plot shows that between 5 and 7 latent topics reside within the dataset.  For this example we could select between 5 and 7 topics to categorize this data.  In this case 5 topics will be selected to continue the analysis.  
 
 ```{r fig.align='center', warning=FALSE, message=FALSE, results='asis'}
[…]
