+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: saotd
 Type: Package
 Title: Sentiment Analysis of Twitter Data
 Version: 0.1.0
-Date: 2018-10-21
+Date: 2018-12-16
 Author: Evan L. Munson 
 Maintainer: Evan L. Munson <evan.l.munson@gmail.com>
 Description: This analytic is an in initial foray into sentiment analysis.  This analytic will allow a user to access the Twitter API (once they create their own developer account), ingest tweets of their interest, clean / tidy data, perform topic modeling if interested, compute sentiment scores utilizing the Bing Lexicon, and output visualizations.
