### 2019-01-02 #764 @evan-l-munson
Gentlemen, good morning.  I have corrected rcpp_hello_world() function.  I was running through everything to make sure it was working properly before I gave you the final word.  Everything was looking good until I tried to view the vignette.  For some reason, the vignette is not found when using utils::vignette("saotd"). I will troubleshoot that and hopefully get that working.  Thanks!
