### 2019-01-20 #764 @evan-l-munson
Good morning, I appreciate your patience with me.  Finishing up this package is ending up being more challenging and time-consuming than I anticipated.  I was working on my vignette issue yesterday and am struggling to fix what I am seeing.  Everything seems to be built properly, however, after I re-download and load the package from my Git page the vignettes are not found.  I have used both, `utils::vignette('saotd')` and `utils::browseVignettes('saotd')` but receive the same error message:  `No vignettes found by utils::browseVignettes("saotd")`.  I have also sent the Git repository to a friend, who experienced the same issue.  As I look at the R package structure I think I have everything correct (and I have been looking at it for months, trying to correct) but obviously I have something incorrect since they are not being found.  If you have some insight as I what might be happening I would appreciate the any thoughts you might have so I can complete this.  Thanks!
