### 2019-02-10 #764 @arfon
> If you have some insight as I what might be happening I would appreciate the any thoughts you might have so I can complete this. Thanks!

I'm not sure sorry. Perhaps @kbenoit has some thoughts on this?
