+++ doc/saotd.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "saotd"
+title: "Sentiment Analysis of Twitter Data"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
