### 2019-02-10 #764 @evan-l-munson
@arfon I have another colleague looking into the issue for me.  I am hoping they will get back to me this week.  I have ran out of ideas on my end and am not sure why I can view the vignettes after I re-download the package from GitHub.  If the vignette issue isn't a big one I would say the package is ready for submission, but if that is a big issue, I will continue to work on it.  Thanks!
