### 2019-02-10 #764 @kbenoit
@evan-l-munson I just tried building your package and it fails with

```r
> devtools::build_vignettes()
Building saotd vignettes
...
Error : processing vignette 'SAoTD.Rmd' failed with diagnostics:
Failed to locate the ‘weave’ output file (by engine ‘knitr::rmarkdown’) for vignette with name ‘SAoTD’. The following files exist in directory ‘.’: ‘saotd.html’, ‘saotd.R’, ‘SAoTD.Rmd’
Error: processing vignette 'SAoTD.Rmd' failed with diagnostics:
Failed to locate the ‘weave’ output file (by engine ‘knitr::rmarkdown’) for vignette with name ‘SAoTD’. The following files exist in directory ‘.’: ‘saotd.html’, ‘saotd.R’, ‘SAoTD.Rmd’
```
but when I changed the vignette name to `test.Rmd` (all lowercase) it succeeded.  I suggest that you change the vignette filename to lowercase (the title can still be anything you want) and hopefully the problem will be solved. It seems related to https://stackoverflow.com/questions/27338970/vignette-creation-on-package-build-fails-with-the-error-failed-to-locate-the-w.

Also I don't think you need **Rcpp** at all. Suggest you delete `/src`, `R/RcppExports.R`, remove Rcpp from the Imports section of `DESCRIPTION`, remove the line `LinkingTo: Rcpp`, and delete the imports in `R/SAoTD.R`. They are doing nothing for you - I suspect they came from a package starter boilerplate tool.
