### 2019-02-11 b3094be
removed vignettes
