### 2019-02-15 #764 @evan-l-munson
@kbenoit I think I have renamed the vignette but still need to run through some checks.  I'm not sure if the `SAoTD.Rmd` is a holdover from when I changed the name of the package to `saotd`.

I haven't had a chance to work on removing the rcpp items but hope to get to that later this weekend.

Again thanks for the help!
