### 2019-02-16 cb3cd85
removed and reinstalled "igraph to 1.2.4"
