### 2019-02-16 bb3d0fe
appveyor.yml r version from release to stable
