+++ .travis.yml
@@ -22,6 +22,7 @@ apt_packages:
 r_binary_packages:
  - topicmodels # for ldatuning.  Insight from:  https://github.com/nikita-moor/ldatuning/blob/master/.travis.yml
  - rmpfr # for ldatuning.  Insight from:  https://github.com/nikita-moor/ldatuning/blob/master/.travis.yml
+ - igraph
  
 r_packages:
  - covr
