+++ appveyor.yml
@@ -27,7 +27,6 @@ test_script:
 on_failure:
 
   - 7z a failure.zip *.Rcheck\*
-
   - appveyor PushArtifact failure.zip
 
 artifacts:
