### 2019-02-18 dfef989
fixed histogram functions, trying to fix appveyor yml
