+++ appveyor.yml
@@ -19,6 +19,7 @@ environment:
 build_script:
 
   - travis-tool.sh install_deps
+  - travis-tool.sh r_binary_install curl
 
 test_script:
 
