### 2019-02-18 39c6528
removed testthat::context() from word corr and corr network
