+++ tests/testthat/test_word_corr.R
@@ -1,4 +1,3 @@
-testthat::context("word_corr")
 
 # Test Data
 
