+++ tests/testthat/test_word_corr_network.R
@@ -1,4 +1,3 @@
-testthat::context("word_corr_network")
 
 # Test Data
 
