### 2019-02-18 54534eb
working CI test failed issues with word_corr and word_corr_diagram
