### 2019-02-18 7fb77b1
added testthat::context() to all tests
