### 2019-02-18 f07e85d
removed igraph from r_binary_packages in travis yml
