### 2019-02-18 8c2b9b8
minor error in word_corr function, define dplyr::n()
