+++ R/word_corr.R
@@ -7,7 +7,7 @@
 #' @param number The number of word instances to be included.
 #' @param sort Rank order the results from most to least correlated.
 #' 
-#' @importFrom dplyr group_by filter quo
+#' @importFrom dplyr group_by filter quo n
 #' @importFrom widyr pairwise_cor
 #' 
 #' @return A tribble
@@ -43,7 +43,7 @@ word_corr <- function(DataFrameTidy, number, sort = TRUE) {
   
   TD_Word_Correlation <- DataFrameTidy %>%
     dplyr::group_by(Token) %>%
-    dplyr::filter(n() >= number) %>%
+    dplyr::filter(dplyr::n() >= number) %>%
     widyr::pairwise_cor(Token, key, sort = sort)
   return(TD_Word_Correlation)
 }
