+++ NAMESPACE
@@ -33,6 +33,7 @@ importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
 importFrom(dplyr,inner_join)
 importFrom(dplyr,mutate)
+importFrom(dplyr,n)
 importFrom(dplyr,quo)
 importFrom(dplyr,select)
 importFrom(dplyr,summarize)
