### 2019-02-19 687fe42
undid all rcpp removals which completely broke the package :-(
