### 2019-02-20 39f0f97
fixed small error in vignette
