+++ vignettes/saotd.R
@@ -1,0 +1,208 @@
+## ----warning=FALSE, message=FALSE----------------------------------------
+library(saotd)
+library(tidytext)
+library(dplyr)
+library(stringr)
+library(knitr)
+library(utils)
+
+## ----eval=FALSE----------------------------------------------------------
+#  consumer_key <- "XXXXXXXXXXXXXXXXXXXXXXXXX"
+#  consumer_secret <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+#  access_token <- "XXXXXXXXXXXXXXXXXX-XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+#  access_secret <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+#  
+#  hashtags <- c("#job", "#Friday", "#fail", "#icecream", "#random", "#kitten", "#airline")
+#  
+#  tweet_acquire(consumer_key = consumer_key,
+#          consumer_secret = consumer_secret,
+#          access_token = access_token,
+#          access_secret = access_secret,
+#          HT = hashtags,
+#          num_tweets = 1000,
+#          file_name = "test_tweets.RData",
+#          distinct = TRUE)
+#  
+#  load("test_tweets.RData")
+
+## ------------------------------------------------------------------------
+data("raw_tweets")
+TD <- raw_tweets
+
+## ------------------------------------------------------------------------
+TD_Tidy <- saotd::tweet_tidy(DataFrame = TD)
+
+TD_Tidy$Token[3:8] %>% 
+  knitr::kable("html")
+
+## ----message=FALSE-------------------------------------------------------
+saotd::unigram(DataFrame = TD) %>% 
+  dplyr::top_n(10) %>% 
+  knitr::kable("html", caption = "Twitter data Uni-Grams")
+
+## ----message=FALSE-------------------------------------------------------
+saotd::bigram(DataFrame = TD) %>% 
+  dplyr::top_n(10) %>% 
+  knitr::kable("html", caption = "Twitter data Bi-Grams")
+
+## ----message=FALSE-------------------------------------------------------
+saotd::trigram(DataFrame = TD) %>% 
+  dplyr::top_n(10) %>% 
+  knitr::kable("html", caption = "Twitter data Tri-Grams")
+
+## ----message=FALSE, message=FALSE, error=FALSE---------------------------
+TD_Merge <- merge_terms(DataFrame = TD, term = "cancelled flight", term_replacement = "cancelled_flight")
+
+## ----message=FALSE-------------------------------------------------------
+saotd::unigram(DataFrame = TD_Merge) %>% 
+  dplyr::top_n(10) %>% 
+  knitr::kable("html", caption = "Twitter data Uni-Grams")
+
+## ----message=FALSE-------------------------------------------------------
+saotd::bigram(DataFrame = TD_Merge) %>% 
+  dplyr::top_n(10) %>% 
+  knitr::kable("html", caption = "Twitter data Bi-Grams")
+
+## ----message=FALSE-------------------------------------------------------
+saotd::trigram(DataFrame = TD_Merge) %>% 
+  dplyr::top_n(10) %>% 
+  knitr::kable("html", caption = "Twitter data Tri-Grams")
+
+## ----fig.align='center'--------------------------------------------------
+TD_Bigram <- saotd::bigram(DataFrame = TD_Merge)
+
+saotd::bigram_network(BiGramDataFrame = TD_Bigram,
+                      number = 90,
+                      layout = "fr",
+                      edge_color = "blue",
+                      node_color = "black",
+                      node_size = 3,
+                      set_seed = 1234)
+
[…]
