### 2019-02-23 #764 @evan-l-munson
@kbenoit  I fixed the vignette issue finally, which of course took longer than I expected.  I went through and removed all Rcpp items in the package and ended up breaking everything.  This weekend, I added all the Rcpp items back into the package, and now the package is passing all tests again and working properly.
