### 2019-02-23 9ccf473
rebuilt after adding rcpp code back in
