+++ man/SAoTD.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/saotd.R
+% Please edit documentation in R/SAoTD.R
 \docType{package}
 \name{saotd}
 \alias{saotd}
