### 2019-02-24 #764 @kbenoit
There is no C++ in the package, so the "breaking" parts were probably due to incomplete removal or failure to rebuild the parts fully when you roxygenized.
