### 2019-02-24 b25183f
Removes the unnecessary Rcpp calls
Fixes the problems as per https://github.com/openjournals/joss-reviews/issues/764#issuecomment-466739716
