+++ src/RcppExports.cpp
@@ -1,27 +1,0 @@
-// Generated by using Rcpp::compileAttributes() -> do not edit by hand
-// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
-#include <Rcpp.h>
-
-using namespace Rcpp;
-
-// rcpp_saotd
-List rcpp_saotd();
-RcppExport SEXP _saotd_rcpp_saotd() {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    rcpp_result_gen = Rcpp::wrap(rcpp_saotd());
-    return rcpp_result_gen;
-END_RCPP
-}
-
-static const R_CallMethodDef CallEntries[] = {
-    {"_saotd_rcpp_saotd", (DL_FUNC) &_saotd_rcpp_saotd, 0},
-    {NULL, NULL, 0}
-};
-
-RcppExport void R_init_saotd(DllInfo *dll) {
-    R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
-    R_useDynamicSymbols(dll, FALSE);
-}
