+++ DESCRIPTION
@@ -7,7 +7,7 @@ Author: Evan L. Munson
 Maintainer: Evan L. Munson <evan.l.munson@gmail.com>
 Description: This analytic is an in initial foray into sentiment analysis.  This analytic will allow a user to access the Twitter API (once they create their own developer account), ingest tweets of their interest, clean / tidy data, perform topic modeling if interested, compute sentiment scores utilizing the Bing Lexicon, and output visualizations.
 License: GPL (>= 2)
-Imports: Rcpp (>= 0.12.14), 
+Imports: 
     plyr, 
     dplyr, 
     widyr,
@@ -28,7 +28,6 @@ Imports: Rcpp (>= 0.12.14),
     magrittr, 
     ldatuning, 
     topicmodels
-LinkingTo: Rcpp
 RoxygenNote: 6.1.1
 Suggests: testthat,
     knitr,
