+++ NAMESPACE
@@ -24,7 +24,6 @@ export(word_corr)
 export(word_corr_network)
 import(ggplot2)
 import(maps)
-importFrom(Rcpp,evalCpp)
 importFrom(dplyr,anti_join)
 importFrom(dplyr,arrange)
 importFrom(dplyr,count)
@@ -69,4 +68,3 @@ importFrom(twitteR,setup_twitter_oauth)
 importFrom(twitteR,twListToDF)
 importFrom(utils,head)
 importFrom(widyr,pairwise_cor)
-useDynLib(saotd, .registration = TRUE)
