### 2019-02-24 5772a87
rebuilt after @kbenoit merge request to fix rcpp issues
