### 2019-02-24 #764 @evan-l-munson
@kbenoit Again thank you for the help with the rcpp items.  I have merged your correction into the master repository.  I retested everything and your corrections have passed all checks.  I think this should be good to go.  Thank you again for your help!
