### 2019-02-24 #764 @evan-l-munson
@arfon with a couple critical pointers from @kbenoit I think that I have finally corrected the last couple of items of concern he had for my package.  Is there anything else that you need from me with this package submission?  Thanks!
