+++ paper.md
@@ -27,9 +27,13 @@ affiliations:
   name: Air Force Institure of Technology
 ---
 
-saotd is an R interface to the Twitter API and can be used to acquire tweets based on user selected \#hashtags.  The package will clean and tidy the Twitter data, determine the latent topics within the tweets utilizing Latent Dirichlet Allocation (LDA), determine a sentiment score using the Bing lexicon dictionary and output visualizations.
+`saotd` is an R package that provides a programmatic interface to the Twitter API and can be used to acquire tweets based on user-specified \#hashtags. The package will clean and tidy the Twitter data, determine the latent topics within the tweets utilizing Latent Dirichlet Allocation (LDA), determine a sentiment score using the Bing lexicon dictionary, and create output visualizations.
 
-The package is available on [GitHub](https://github.com/evan-l-munson/saotd) and archived on [Zenodo](https://zenodo.org/record/1219852#.WtdNuchryfd).  The package was developed with the intention of the user creating a personal [Twitter](hjttps://twitter.com) account which will then allow  a user to access the twitter API throught the [Twitter Developers Account](https://dev.twitter.com/) site.  Once a user has access to the Twitter Developers Account they will have the ability to create an application which will then provide the user with access tokens.  These access tokens will then allow the user to begin acquiring tweets usint eh saotd package.
+The package is available on [GitHub](https://github.com/evan-l-munson/saotd) and archived on [Zenodo](https://zenodo.org/record/1219852#.WtdNuchryfd). To configure the package a user must follow these steps:
+
+* Creating a personal [Twitter](https://twitter.com) account if they don't already have one.
+* Enable a [Twitter Developers Account](https://dev.twitter.com/).
+* Create an application which will provide them with API access tokens required by the `saotd` package to function.
 
 The package is laid out in five different categories: Acquire, Explore, Topic Analysis, Sentiment Calculation, and Visualizations.
 
@@ -37,11 +41,11 @@ The package is laid out in five different categories: Acquire, Explore, Topic An
 * Explore provides functions to tidy, explore unigrams, bigrams, tri-grams, in addition to bigram netwrorks and correlation networks.
 * Topic analysis allows a user to explore the latent topics buried within the tweets.
 * Sentiment calculation utilizes the Bing lexicon dictionary to score the text [@Hu2004].
-* Visualizations allow the user to better understand the twitter sentiment.
+* Visualizations allow the user to better understand the sentiment of the tweets.
 
-The package utilizes tidy dataferames and therefore depends on the tidyverse package [@Wickham2017] and additionally uses the tidytext package [@Silge2017].  The number of latent topics is determined using the ldatuning package [@Nikita2016] and the latent dirichlet allocation (LDA) topics is determined using the topicmodels package [@Grun2011].
+The package utilizes tidy dataframes and therefore depends on the `tidyverse` package [@Wickham2017] and additionally uses the `tidytext` package [@Silge2017].  The number of latent topics is determined using the `ldatuning` package [@Nikita2016] and the latent dirichlet allocation (LDA) topics is determined using the `topicmodels` package [@Grun2011].
 
-The saotd package has research applications in many disciplines which intend to use twitter text and sentiment analysis.  The package was created to quickly determine the sentiment of twitter and to inform analysts on the opinions contained within tweets.
+The `saotd` package has research applications in many disciplines which need to access tweets from the Twitter platform and carry out sentiment analyses. This package was created to quickly determine the sentiment of Twitter and to inform analysts on the opinions contained within tweets.
 
 # References
 
