### 2019-02-25 #764 @arfon
Hi @evan-l-munson. I've made some slight tweaks to your paper in https://github.com/evan-l-munson/saotd/pull/9 - let me know what you think.

In addition, please make sure you:

- Update the version number based on the feedback from this review.
- Add something to your README making it clear how users can contribute (if appropriate).
