### 2019-02-25 064dd79
minor description updates and version bump to 0.2.0
