### 2019-02-25 #764 @evan-l-munson
@arfon, Good evening.  I have completed your suggested edits/additions and pushed to GitHub.  If there is anything else, let me know and I will get working on it asap.  Thanks for the help!
