### 2019-02-25 #764 @arfon
> @arfon, Good evening. I have completed your suggested edits/additions and pushed to GitHub. If there is anything else, let me know and I will get working on it asap. Thanks for the help!

@evan-l-munson thanks. Please could you clarify what version the package is now at? Given the modifications during the review it would seem appropriate to make a new release and archive in Zenodo.

Once you've done this I'm happy to proceed with accepting this submission.
