### 2019-02-25 #764 @evan-l-munson
@arfon, I bumped the package to version 0.2.0.  I will work to get resubmit to Zendo within the next day or two.  Will let you know when I get that done. Thanks!
