### 2019-02-25 7a4945b
correcting cran url errors
