+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: saotd
 Type: Package
 Title: Sentiment Analysis of Twitter Data
 Version: 0.2.0
-Date: 2019-02-24
+Date: 2019-02-26
 Authors@R: c(
     person("Evan", "Munson", 
            email = "evan.l.munson@gmail.com", 
@@ -21,9 +21,9 @@ Authors@R: c(
            role = c("aut"),
            comment = c(ORCID = "0000-0002-2415-0340"))
            )
-Maintainer: Evan L. Munson <evan.l.munson@gmail.com>
+Maintainer: Evan Munson <evan.l.munson@gmail.com>
 BugReports: https://github.com/evan-l-munson/saotd
-Description: This analytic is an in initial foray into sentiment analysis.  This analytic will allow a user to access the Twitter API (once they create their own developer account), ingest tweets of their interest, clean / tidy data, perform topic modeling if interested, compute sentiment scores utilizing the Bing Lexicon, and output visualizations.
+Description: This analytic is an in initial foray into sentiment analysis.  This analytic will allow a user to access the Twitter API (once they create their own developer account), ingest tweets of their interest, clean / tidy data, perform topic modeling if interested, compute sentiment scores utilizing the bing Lexicon, and output visualizations.
 License: GPL (>= 2)
 Imports: 
     plyr, 
