### 2019-02-27 c891007
prepare for cran resubmission and update to zenodo
