+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: saotd
 Type: Package
 Title: Sentiment Analysis of Twitter Data
 Version: 0.2.0
-Date: 2019-02-26
+Date: 2019-02-27
 Authors@R: c(
     person("Evan", "Munson", 
            email = "evan.l.munson@gmail.com", 
