### 2019-02-27 #764 @evan-l-munson
@arfon, I released the 0.2.0 version to zenodo this morning and have updated the DOI badge.  Is there anything else you need?  Thanks!
